// src/generated/assets.ts
/**
 * AUTO-GENERATED FILE - DO NOT EDIT
 * Generated by react-assets-gen
 * Generated at: 2025-12-14T17:36:32.311Z
 */

// SVG React Components
import iconsCloseComponent from '../../assets/icons/close.svg?react';
import iconsInfoComponent from '../../assets/icons/info.svg?react';
import iconsMenuComponent from '../../assets/icons/menu.svg?react';

/**
 * Type-safe assets object
 */
export interface AssetsType {
  readonly icons: {
    readonly close: string;
    readonly info: string;
    readonly menu: string;
  };
  readonly images: {
    readonly logo: string;
  };
}

/**
 * Assets object containing all asset paths
 * @example
 * import { Assets } from "./generated/assets";
 * <img src={Assets.images.logo} alt="Logo" />
 */
export const Assets: AssetsType = {
  icons: {
    close: '../../assets/icons/close.svg',
    info: '../../assets/icons/info.svg',
    menu: '../../assets/icons/menu.svg',
  },
  images: {
    logo: '../../assets/images/logo.png',
  },
} as const;

/**
 * SVG React Components
 * @example
 * import { SvgComponents } from "./generated/assets";
 * <SvgComponents.icons.close />
 */
export const SvgComponents = {
  iconsClose: iconsCloseComponent,
  iconsInfo: iconsInfoComponent,
  iconsMenu: iconsMenuComponent,
} as const;

export default Assets;